Remove hex color preference
There are many useful color notations beyond hexadecimal, for example
`rgba()` when you need alpha and `hsl()`/`hsla()` when you want to use
the HSL cylindrical-coordinate system.

Having this guide to prefer hexadecimal feels a bit limiting and I think
people tend to stick to a consistent convention within a project, which
is more important than the specific notation.